+++ lib/ui/wellness/WellnessSectionsContentWidget.dart
@@ -15,9 +15,13 @@
  */
 
 import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
 import 'package:illinois/service/Analytics.dart';
+import 'package:illinois/service/Auth2.dart';
+import 'package:illinois/service/Transportation.dart';
 import 'package:illinois/ui/wellness/WellnessEightDimensionsPanel.dart';
 import 'package:rokwire_plugin/service/localization.dart';
+import 'package:rokwire_plugin/service/notification_service.dart';
 import 'package:rokwire_plugin/service/styles.dart';
 import 'package:rokwire_plugin/ui/widgets/rounded_button.dart';
 
@@ -28,21 +32,54 @@ class WellnessSectionsContentWidget extends StatefulWidget {
   State<WellnessSectionsContentWidget> createState() => _WellnessSectionsContentWidgetState();
 }
 
-class _WellnessSectionsContentWidgetState extends State<WellnessSectionsContentWidget> {
+class _WellnessSectionsContentWidgetState extends State<WellnessSectionsContentWidget> implements NotificationsListener {
+  Color? _tipColor;
+  bool _loading = false;
+
+  @override
+  void initState() {
+    super.initState();
+    NotificationService().subscribe(this, [Auth2.notifyLoginChanged]);
+    _loadTipColor();
+  }
+
+  @override
+  void dispose() {
+    NotificationService().unsubscribe(this);
+    super.dispose();
+  }
+
+  void _loadTipColor() {
+    _setLoading(true);
+    Transportation().loadBusColor(userId: Auth2().accountId, deviceId: Auth2().deviceId).then((activeColor) {
+      _tipColor = activeColor;
+      _setLoading(false);
+    });
+  }
+
   @override
   Widget build(BuildContext context) {
-    return _buildContent();
+    return _loading ? _buildLoadingContent() : _buildContent();
   }
 
   Widget _buildContent() {
     return Column(
         crossAxisAlignment: CrossAxisAlignment.center,
-        children: [_buildHeaderDescription(), _buildEightDimensionImage(), _buildFooterDescription(), _buildEightDimensionButton()]);
+        children: [_buildTipDescription(), _buildEightDimensionImage(), _buildFooterDescription(), _buildEightDimensionButton()]);
   }
 
-  Widget _buildHeaderDescription() {
+  Widget _buildLoadingContent() {
+    return Center(
+        child: Column(children: <Widget>[
+      Container(height: MediaQuery.of(context).size.height / 5),
+      CircularProgressIndicator(),
+      Container(height: MediaQuery.of(context).size.height / 5 * 3)
+    ]));
+  }
+
+  Widget _buildTipDescription() {
     return Container(
-        color: Styles().colors!.accentColor3,
+        color: (_tipColor ?? Styles().colors!.accentColor3),
         padding: EdgeInsets.all(42),
         child: Text(
             Localization().getStringEx('panel.wellness.sections.description.header.text',
@@ -88,5 +125,21 @@ class _WellnessSectionsContentWidgetState extends State<WellnessSectionsContentW
   void _onTapEightDimensions() {
     Analytics().logSelect(target: "Wellness 8 Dimensions");
     Navigator.of(context).push(CupertinoPageRoute(builder: (context) => WellnessEightDimensionsPanel()));
+  }
+
+  void _setLoading(bool loading) {
+    _loading = loading;
+    if (mounted) {
+      setState(() {});
+    }
+  }
+
+  // Notifications Listener
+
+  @override
+  void onNotification(String name, param) {
+    if (name == Auth2.notifyLoginChanged) {
+      _loadTipColor();
+    }
   }
 }